cummulative changes

+++ pages/auth.tsx
@@ -7,7 +7,7 @@ import { auth, db } from "../.env/firebase";
 import * as firebase from "firebase/compat/app";
 import data from "../public/Major_Names.json";
 import Error from "../components/error";
-import { getStorage, ref, uploadString } from "firebase/storage";
+import { getDownloadURL, getStorage, ref, uploadBytes, uploadBytesResumable, uploadString } from "firebase/storage";
 import { app, firestore } from '../firebase';
 import { collection, addDoc, getDoc, getDocs, query, where } from 'firebase/firestore';
 import GooglePlacesAutocomplete, { getLatLng } from 'react-google-places-autocomplete';
@@ -20,6 +20,7 @@ const dbEmailInstance = collection(firestore, 'emails');
 const Authentication = () => {
   let [err, setError] = React.useState({ active: false, code: 0 , error: true});
   let [pageNo, change] = React.useState(1);
+  let [imageUrl, changeUrl] = React.useState("");
   let [location, selectLocation] = React.useState(null as any)
   let [userInformation, addInfo] = React.useState({
     email: "",
@@ -51,6 +52,21 @@ const Authentication = () => {
       // restore defaults
       setError({ active: err.active, code: code, error:isError });
     }
+  }
+   async function uploadFile(): Promise<string>{
+    const uploadRef = ref(storage, "storage/"+image?.name);
+    
+    const uploadTask =  await uploadBytesResumable(uploadRef, image as any)
+    const imageUrl = await getDownloadURL(uploadTask.ref).then((downloadURL) => {
+        
+          console.log('File available at', downloadURL);
+          changeUrl(downloadURL)
+          userInformation.picture = downloadURL;
+          return downloadURL;
+        }).catch(async (er)=>{
+          await error(2)
+        });
+  return imageUrl!
   }
   function submitEmail(email: string,n:number): boolean {
     // const auth = getAuth(app);
@@ -113,19 +129,22 @@ const Authentication = () => {
     try {
       const res = await fetch(`https://maps.googleapis.com/maps/api/geocode/json?place_id=${placeId}&key=AIzaSyDcjNrNrDamH1BaZ6BtgvWY3ENNx5QXoM4`);
       const obj = await res.json();
+      console.log(obj)
       // var obj = JSON.parse(data);
       let lat =  obj.results[0].geometry.location.lat;
       let lng =  obj.results[0].geometry.location.lng; 
-      
-      addInfo({
+      userInformation.lat = lat;
+      userInformation.lng = lng;
+      await uploadFile();
+      await addInfo({
         email:  window.localStorage.getItem("email") || getInputVal("email"),
         name: getInputVal("name"),
         major: getInputVal("major"),
-        country: getInputVal("country"),
-        city: getInputVal("city"),
-        picture: "",
-        lat: lat,
-        lng: lng
+        city: location?.label.split(',').slice(0,-1),
+        country: location?.label.split(',').slice(-1),
+        picture: imageUrl,
+        lat: obj.results[0].geometry.location.lat,
+        lng: obj.results[0].geometry.location.lng
       });
       console.log(userInformation)
       console.log(image)
@@ -167,9 +186,10 @@ const Authentication = () => {
     const storageRef = ref(storage, "storage");
     const imagesRef = storageRef;
     const message1 = "input-storage";
-    uploadString(storageRef, message1).then((snapshot) => {
-      console.log("Successful upload!");
-    });
+   // uploadString(storageRef, message1).then((snapshot) => {
+        //   console.log(snapshot)
+        //   console.log("Successful upload!");
+        // });
     
    async function next(e: any, n: number) {
     e.preventDefault();
@@ -186,13 +206,33 @@ const Authentication = () => {
       // add if user information is null - use kris function coral
         userInformation.name = getInputVal("name");
         userInformation.major = getInputVal("major");
-        userInformation.country = getInputVal("country");
-        userInformation.city = getInputVal("city");
-        change(n);
+        userInformation.city = location?.label.split(',').slice(0,-1);
+        userInformation.country = location?.label.split(',').slice(-1);
+        userInformation.email = getInputVal("email");
+        if(userInformation.name=="" || userInformation.major=="" || userInformation.city=="" || userInformation.country=="" || userInformation.email==""){
+            
+            if(userInformation.city==""){
+              await error(7)
+            }
+            else{
+              await error(4);
+            }
+        }
+        else{
+           change(n);
+        }
+       
         break;
       case 0:
-        await getLatLng(location.value.place_id)
-        window.localStorage.clear()
+        if(image){
+          await getLatLng(location.value.place_id)
+        
+          window.localStorage.clear()
+        }
+        else{
+          await error(8)
+        }
+       
         break;
       default:
         change(n);
@@ -206,7 +246,7 @@ const Authentication = () => {
  
   return (
     <div className={global.font}>
-      <form>
+      <form className={styles.center_flex}>
         {/* step one: email input */}
         <div className={pageNo == 1 ? styles.form : global.hidden}>
           <h1>Type your student email</h1>
@@ -224,7 +264,9 @@ const Authentication = () => {
         {/* step two: image input */}
         <div className={pageNo == 2 ? styles.form : global.hidden}>
           <h1>add information</h1>
+          <p>Name</p>
           <input type="text" placeholder="Name" name="name" id="name"></input>
+          <p>Major</p>
           <select name="major" id="major">
             {data["MAIN"].map((val)=><option value={val['Major Name']} key={val['Major Name']}>{val["Major Name"]}</option>)}
         </select>
@@ -257,16 +299,8 @@ const Authentication = () => {
                 }}
               />
         </div>
-          
-          <input type="text" placeholder="City" name="city" id="city" value={location?.label.split(',').slice(0,-1)} readOnly></input>
-          <input
-            type="text"
-            placeholder="Country"
-            name="country"
-            id="country"
-            value={location?.label.split(',').slice(-1)} readOnly
-          ></input>
-          {/* submit everything to firebase on this step */}
+          <p>Country Of Origin: <strong>{location?.label.split(',').slice(-1)}</strong></p>
+          <p>City of Origin: <strong>{location?.label.split(',').slice(0,-1)}</strong></p>
           <div  className={styles.bottomBtns}>
             <button className={`${global.button_secondary} ${global.button}`}  onClick={(e) => back(e, 1)}> back </button>
